Chocolatey-Uninstall: fix passing -uninstall switch to Get-ChocolateyBins
This fixes batch redirect files not being removed when uninstalling a
package (a bug introduced by b12e91bf "add error level checking...").

Invoke-ChocolateyFunction uses the splatting operator ("@") to pass
parameters to the invoked function. When the parameter values are passed to
Invoke-ChocolateyFunction encapsulated in an array, the splatting operator
matches them only to positional parameters of the invoked function. In
order to pass values of named parameters and/or switches, they need to be
encapsulated in a hashtable.