tests: remove chocolateyInstaller module at test initialization
The existing test init code removes the 'chocolatey' and '_Setup' modules,
but leaves the 'chocolateyInstaller' module loaded. This has the nasty
effect that changes to chocolatey helpers are ignored when tests are
repeatedly run in the same Powershell session using Invoke-Pester.

As 'chocolateyInstaller' is imported as a nested module (by the
'chocolatey' module), it requires the -All switch to Get-Module in order
to be discovered and must be removed by piping it to Remove-Module.

Also change the existing module removal commands to pipe syntax, so as not
to mix two removal styles in one file (and get tighter code).