Enable mocking Test-AdminRights for setup tests
As Chocolatey setup code explicitly imports the helpers module, Pester
mocking mechanisms do not work (the mocks are overridden when the helpers
are imported again).

Mocking the admin rights check in setup tests is critical to making it
possible to test Chocolatey setup at various user right levels. The only
way this developer has found to achieve that was to implement a test hook
in the helper itself.