remove test hook from Test-AdminRights
Do not taint production code with test hooks - remove mocking ability from
regular implementation of Test-AdminRights.

For setup tests, substitute Test-AdminRights implementation in the
installation package with a stub returning a canned answer.

This slightly reduces the accuracy of setup tests, as not 100% of
production code is tested, but keeping production code pure is more
important.

As an added benefit, this is the first step in enabling the tests to be
run as a standard user.