use environment mocking facility in Update-SessionEnvironment tests
This simplifies and clarifies test setup code and decouples it from
implementation details, not to mention fixing the cases broken after the
helper started using the environment manipulation functions.

These tests can now be run as standard user, eliminating the fear of
corrupting systemwide environment in case of a bug.